Changes to be committed: 	modified:   app/controllers/search_controller.rb 	modified:   app/helpers/search_helper.rb 	modified:   app/views/search/_search_sidebar.html.erb
 Changes not staged for commit:
	modified:   Gemfile
	modified:   Gemfile.lock

 Untracked files:
	.gitignore
	.idea/

Workign on getting the sidbar that handles the faceting to work.  The main files for this are search_helper.rb, which contains all the functional code, and _search_sidebar.html.erb, which controls the view.  The call to the function that creates the faceting is currently at line 40 of _search_sidebar.html.erb.  This calls create_genre_table() at line 522 of search_helper.rb.  Right now this is successfully loading the 'genre' facet, which contains format information in the new schema.  I've also added some code to create a temporary new facet with hard-coded info.  This is where I'll work to creat the actual facets.  The plan here is to use a "see all" link to go get the facets via Javascript when it is clicked instead of having to load it all upfront.