Changes to be committed: 	modified:   app/helpers/search_helper.rb 	modified:   app/views/search/_search_sidebar.html.erb
Cleaned up the building of facet blocks in the left hand sidebar.  Now _search_sidebar.html.erb calls new, separate functions in search_helper.rb for each facet.  (right now there is only the original genre facet and a test one that does nothing.)  The way the code works, in all cases search_helper.rb builds an empty table (just a <th> row) and the contents gets filled in by the jscript at Line 32 of /assets/searh_results_draw_facets.js in the function createFacetBlock().  Jquery is currently doing this for the 'genre' facet at lines 180, 181, 182, and 183 of searh_results_draw_facets.js.

Next move is to determine the final format of the facet blocks, and then modify the createFacetBlock() so that it is only called when some clicks on the 'see more' link.